+++ test/msw/handlers.ts
@@ -1,0 +1,4 @@
+import { RestHandler } from 'msw';
+import { scheduledJobHandlers } from './handlers/scheduledJobs';
+
+export const handlers: RestHandler[] = [...scheduledJobHandlers];