+++ test/msw/server.ts
@@ -1,0 +1,6 @@
+import { SetupServer, setupServer } from 'msw/node';
+import { handlers } from './handlers';
+
+const server: SetupServer = setupServer(...handlers);
+
+export default server;